mes-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.43
pydantic
pydantic-settings
psycopg2-binary
python-dotenv==1.0.0
